Refactor displayAddress to return JSX
