+++ apps/modernization-ui/src/address/display/displayAddress.scss
@@ -1,0 +1,8 @@
+address {
+    margin-bottom: 0.5em;
+}
+
+address > label {
+    display: block;
+    font-weight: bold;
+}
