--- apps/modernization-ui/src/address/display/displayAddress.ts
@@ -1,18 +1,0 @@
-import { exists } from 'utils';
-
-export type DisplayableAddress = {
-    address?: string | null;
-    address2?: string | null;
-    city?: string | null;
-    state?: string | null;
-    zipcode?: string | null;
-};
-
-const displayAddress = ({ address, address2, city, state, zipcode }: DisplayableAddress) => {
-    const street = [address, address2].filter(exists).join(', ');
-    let location = [city, state].filter(exists).join(', ');
-    location = [location, zipcode].filter(exists).join(' ');
-    return [street, location].filter(exists).join('\n');
-};
-
-export { displayAddress };
